Rename package to fix sonar warning - all lowercased now
