+++ apache-rat-core/src/main/java/org/apache/rat/config/exclusion/StandardCollection.java
@@ -29,11 +29,11 @@ import java.util.Set;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 
-import org.apache.rat.config.exclusion.fileProcessors.AbstractFileProcessorBuilder;
-import org.apache.rat.config.exclusion.fileProcessors.BazaarIgnoreBuilder;
-import org.apache.rat.config.exclusion.fileProcessors.CVSIgnoreBuilder;
-import org.apache.rat.config.exclusion.fileProcessors.GitIgnoreBuilder;
-import org.apache.rat.config.exclusion.fileProcessors.HgIgnoreBuilder;
+import org.apache.rat.config.exclusion.fileprocessors.AbstractFileProcessorBuilder;
+import org.apache.rat.config.exclusion.fileprocessors.BazaarIgnoreBuilder;
+import org.apache.rat.config.exclusion.fileprocessors.CVSIgnoreBuilder;
+import org.apache.rat.config.exclusion.fileprocessors.GitIgnoreBuilder;
+import org.apache.rat.config.exclusion.fileprocessors.HgIgnoreBuilder;
 import org.apache.rat.document.DocumentName;
 import org.apache.rat.document.DocumentNameMatcher;
 import org.apache.rat.utils.ExtendedIterator;
