+++ apache-rat-core/src/test/java/org/apache/rat/config/exclusion/fileprocessors/AbstractIgnoreBuilderTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.rat.config.exclusion.fileProcessors;
+package org.apache.rat.config.exclusion.fileprocessors;
 
 import java.nio.file.Path;
 import java.util.List;
