+++ apache-rat-plugin/spotbugs-ignore.xml
@@ -19,7 +19,7 @@
   <Match>
     <!--
     Convenience constructors that allow setting a charset are not available in Java8 for FileReader.
-    <Class name="org.apache.rat.config.exclusion.fileProcessors.GlobIgnoreMatcher"/>
+    <Class name="org.apache.rat.config.exclusion.fileprocessors.GlobIgnoreMatcher"/>
     <Bug pattern="DM_DEFAULT_ENCODING"/>
   -->
   </Match>
